create and flesh out admin namespace
