+++ app/controllers/admin/visits_controller.rb
@@ -1,7 +1,4 @@
-class VisitsController < ApplicationController
-  
-  before_filter :require_authentication
-  
+class Admin::VisitsController < AdminController
   def index
     @short = Short.find(params[:short_id])
     @visits = @short.visits
@@ -9,5 +6,4 @@ class VisitsController < ApplicationController
     @clicks_per_day = @short.visits.group("DATE(created_at)").order("created_at DESC").limit(30).count
     @referrers = @short.visits.group("referred").order("count_all DESC").limit(10).count
   end
-  
 end
