+++ app/controllers/application_controller.rb
@@ -7,10 +7,16 @@ class ApplicationController < ActionController::Base
 
     def current_user
       @current_user ||= User.find(session[:user_id]) if session[:user_id]
+      @current_user
     end
 
     def require_authentication
-      redirect_to "/auth/google_oauth2" unless current_user
+      # don't worry about logging in while developing
+      if Rails.env == "development"
+        @current_user ||= User.first
+      else
+        redirect_to "/auth/google_oauth2" unless current_user
+      end
     end
 
     def render_404
