+++ app/helpers/application_helper.rb
@@ -1,11 +1,9 @@
 module ApplicationHelper
-  
   def short_url(short)
     "#{root_url}#{short.contracted}"
   end
-  
+
   def stats_url(short)
-    "#{root_url}#{short.contracted}+"
+    "#{admin_root_url}/#{short.contracted}"
   end
-  
 end
