+++ app/views/admin/shorts/edit.html.erb
