+++ app/views/admin/shorts/index.html.erb
