+++ app/views/admin/shorts/new.html.erb
