+++ app/views/admin/users/index.html.erb
@@ -16,9 +16,9 @@
 	<% @user.shorts.each do |short| %>
 		<tr>
 			<td><%= link_to short.title, stats_url(short) %></td>
-			<td><%= link_to(short_url(short), short.contracted, target: "_blank") %></td>
+			<td><%= link_to(short.contracted, short_url(short), target: "_blank") %></td>
 			<td><%= short.visits.count %></td>
-			<td><%= link_to "edit", edit_short_path(short) %></td>
+			<td><%= link_to "edit", edit_admin_short_path(short) %></td>
 		</tr>
 	<% end %>
 	</tbody>
