+++ app/views/admin/visits/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_for(@visit) do |f| %>
+<%= form_for([:admin, @visit]) do |f| %>
   <% if @visit.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(@visit.errors.count, "error") %> prohibited this visit from being saved:</h2>
