+++ app/views/admin/visits/edit.html.erb
@@ -3,4 +3,4 @@
 <%= render 'form' %>
 
 <%= link_to 'Show', @visit %> |
-<%= link_to 'Back', visits_path %>
+<%= link_to 'Back', admin_visits_path %>
