+++ app/views/admin/visits/index.html.erb
@@ -1,4 +1,4 @@
-<h1><%= short_url @short %> <span class="small"><%= link_to "edit", edit_short_path(@short) %></span></h1>
+<h1><%= short_url @short %> <span class="small"><%= link_to "edit", edit_admin_short_path(@short) %></span></h1>
 
 <div id="short-stats">
 	<h2><%= @short.title %></h2>
