+++ app/views/admin/visits/new.html.erb
@@ -2,4 +2,4 @@
 
 <%= render 'form' %>
 
-<%= link_to 'Back', visits_path %>
+<%= link_to 'Back', admin_visits_path %>
