+++ app/views/admin/visits/show.html.erb
@@ -11,5 +11,5 @@
 </p>
 
 
-<%= link_to 'Edit', edit_visit_path(@visit) %> |
-<%= link_to 'Back', visits_path %>
+<%= link_to 'Edit', edit_admin_visit_path(@visit) %> |
+<%= link_to 'Back', admin_visits_path %>
