+++ app/views/layouts/application.html.erb
@@ -1,4 +1,4 @@
-<!DOCTYPE html> 
+<!DOCTYPE html>
 <html>
 <head>
 	<title><%= @page_title || "Shorten that URL" %> | bcrr.us</title>
@@ -11,14 +11,16 @@
 	<div class="container">
 		<header>
 			<div class="span-5" id="logo">
-				<%= link_to(image_tag("logo.png"), root_path)%>
+				<%= link_to(image_tag("logo.png"), admin_root_path)%>
 			</div>
 			<nav class="span-19 last">
 				<ul>
-					<% if current_user %>
-						<li><%= link_to "My Account", users_path %></li>
-						<li><%= link_to "Site Keys", new_user_path %></li>
+					<% if current_user.present? %>
+						<li><%= link_to "My Account", admin_users_path %></li>
+						<li><%= link_to "Site Keys", new_admin_user_path %></li>
 						<li><%= link_to "Sign Out", signout_path %></li>
+					<% else %>
+						<li>No User!</li>
 					<% end %>
 				</ul>
 			</nav>
