+++ config/routes.rb
@@ -4,16 +4,20 @@ Shorty::Application.routes.draw do
   match "/auth/failure" => "sessions#failure", via: :get
   match "/signout" => "sessions#destroy", :as => :signout, via: :get
 
-  resources :shorts do
-    resources :visits
+  namespace :admin do
+    resources :shorts do
+      resources :visits
+    end
+
+    resources :users
   end
 
   namespace :api do
     resources :shorts
   end
 
-  resources :users
-
-  root :to => "shorts#index"
+  root :to => redirect("http://bethel.global/")
+  match '/admin/', :to => "admin/shorts#index", via: :get, :as => :admin_root
+  match '/admin/*a', :to => 'admin/shorts#show', via: :get
   match '*a', :to => 'shorts#show', via: :get
 end
