+++ spec/controllers/admin/visits_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe VisitsController do
+describe Admin::VisitsController do
   context "access control" do
     it "requires user to be present" do
       short = Short.create!(expanded: "http://yahoo.com")
