--- spec/controllers/users_controller_spec.rb
@@ -1,12 +1,0 @@
-require 'spec_helper'
-
-describe UsersController do
-
-  describe "GET 'index'" do
-    it "should be successful" do
-      get 'index'
-      expect(response).to be_successful
-    end
-  end
-
-end
